Updated ReactionManager tests.
