+++ src/test/java/com/github/splendor_mobile_game/websocket/handlers/reactions/test_classes/TestClassWithoutInterface.java
@@ -1,0 +1,5 @@
+package com.github.splendor_mobile_game.websocket.handlers.reactions.test_classes;
+
+public class TestClassWithoutInterface {
+
+}
